Manambatra projet

+++ src/main/java/com/enchere/enchere/repository/ChiffreAffaireRepository.java
@@ -35,7 +35,7 @@ public class ChiffreAffaireRepository {
                 new BeanPropertyRowMapper<Produit>(Produit.class));
         System.out.println(produits.size());
         for (int i = 0;i < produits.size(); i++) {
-            float prix = ((Produit)produits.get(i)).getPrix();
+            double prix = ((Produit)produits.get(i)).getPrix();
             int id = ((Produit)produits.get(i)).getId();
             String insert = "INSERT INTO Commission VALUES (DEFAULT,((SELECT taux FROM TauxComission ORDER BY date DESC LIMIT 1)/100)*("+prix+"),CURRENT_TIMESTAMP,?)";
             String update = "UPDATE Produit set etat = 1 WHERE id = ?";
